feat: enhance camera system with WASD movement, add raymarching libraries and improve templates
- Added 'u_fork_cam_pos' uniform and implemented WASD flying/orbital controls in C++.
- Added camera position persistence in '.4k-eater.local'.
- Created 'raymarching.glsl' and 'raymarching_vec2.glsl' libraries.
- Updated 'camera.glsl' with proper FOV/focal length and WASD support.
- Updated '#pragma slider' to support floating-point values.
- Enhanced all shader templates with library includes and interactive sliders.
- Improved 'ShaderPreprocessor' to search for includes in 'libs/' directory.
- Fixed syntax errors and added utility functions to 'color.glsl'.
- New projects now automatically export bundled libraries.

+++ src/ParameterPanel.cpp
@@ -126,9 +126,9 @@ void ParameterPanel::render(const std::string& shaderName) {
             }
 
             std::string label = sl.label.empty() ? sl.name : sl.label;
-            int value = m_shaderManager->getSliderState(sl.name);
+            float value = m_shaderManager->getSliderState(sl.name);
             
-            if (ImGui::SliderInt(label.c_str(), &value, sl.min, sl.max)) {
+            if (ImGui::SliderFloat(label.c_str(), &value, sl.min, sl.max)) {
                 m_shaderManager->setSliderState(sl.name, value);
                 RenderScaleMode scaleMode = Settings::getInstance().getRenderScaleMode();
                 if (m_shaderManager->reloadShader(shaderName, scaleMode)) {
